requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-stdimage==6.0.2
pytils==0.4.1

django-debug-toolbar==4.2.0
psycopg2-binary==2.9.9
